requirements.txt
numpy==1.23.4
gym==0.26.2
scipy==1.9.3
matplotlib==3.5.3
pyyaml==6.0
tqdm==4.64.1
opencv-python==4.7.0.72
torch==1.13.0
tensorboard==2.12.0
